chore(mpt): mpt noop trait impls
